test: extract MockCommandGroupRepository
