+++ internal/commandgroup/application/usecases/createcommandgroup_test.go
@@ -17,7 +17,7 @@ import (
 func TestDefaultCreateCommandGroup_Execute(t *testing.T) {
 	t.Run("Should create a command group", func(t *testing.T) {
 		// Arrange
-		mockCommandGroupRepository := new(MockCommandGroupRepository)
+		mockCommandGroupRepository := new(test2.MockCommandGroupRepository)
 		mockUserConfigRepository := new(MockConfigRepository)
 
 		projectId := "project1"
@@ -59,7 +59,7 @@ func TestDefaultCreateCommandGroup_Execute(t *testing.T) {
 
 	t.Run("Should return an error if failing to retrieve user config", func(t *testing.T) {
 		// Arrange
-		mockCommandGroupRepository := new(MockCommandGroupRepository)
+		mockCommandGroupRepository := new(test2.MockCommandGroupRepository)
 		mockUserConfigRepository := new(MockConfigRepository)
 
 		sut := usecases.NewCreateCommandGroup(mockUserConfigRepository, mockCommandGroupRepository)
@@ -85,7 +85,7 @@ func TestDefaultCreateCommandGroup_Execute(t *testing.T) {
 
 	t.Run("Should return an error if failing to retrieve existing command groups", func(t *testing.T) {
 		// Arrange
-		mockCommandGroupRepository := new(MockCommandGroupRepository)
+		mockCommandGroupRepository := new(test2.MockCommandGroupRepository)
 		mockUserConfigRepository := new(MockConfigRepository)
 
 		projectId := "project1"
@@ -114,7 +114,7 @@ func TestDefaultCreateCommandGroup_Execute(t *testing.T) {
 
 	t.Run("Should return an error if failing to save the command group", func(t *testing.T) {
 		// Arrange
-		mockCommandGroupRepository := new(MockCommandGroupRepository)
+		mockCommandGroupRepository := new(test2.MockCommandGroupRepository)
 		mockUserConfigRepository := new(MockConfigRepository)
 
 		projectId := "project1"
